Add ProductoExoneracion entity and exonerated product support
New ProductoExoneracion entity and ProductoExoneracionService for tracking product-level tax exonerations (reduccion, exoneracion, exemption). ArticulosController updated with exoneracion management. Articulos and ArticuloCarrito models extended to support exoneracion lookups.

+++ mercurius-quarkus/src/main/java/Models/Articulos/Carrito/ArticuloCarrito.java
@@ -99,7 +99,7 @@ public class ArticuloCarrito {
         return CarritoCalculations.calculateTotalCarrito(carrito);
     }
 
-    public static Map<Integer, BigDecimal> calculateTotalTaxForUniqueRates(List<ArticuloCarrito> carrito) {
+    public static Map<BigDecimal, BigDecimal> calculateTotalTaxForUniqueRates(List<ArticuloCarrito> carrito) {
         return CarritoCalculations.calculateTotalTaxByRate(carrito);
     }
